requirements-gpu.txt
bleach==1.5.0
click==6.7
cycler==0.10.0
enum34==1.1.6
Flask==0.12.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
h5py==2.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
-e git://github.com/fchollet/keras.git@eac78b859beb31cafa65a3edb4eaa888d3b6c2e6#egg=Keras
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.1
numpy==1.13.3
opencv-python==3.4.0.12
Pillow==5.0.0
protobuf==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.2.1
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.1
